# BackEnd/blog/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "version": "1.0.0"
    },
    "host": "localhost:800/hydroponic/BackEnd/blog/public",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/devices/getDeviceIdForUser": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "get data",
                "description": "get data",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/devices/getSensorData": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "get sensor data",
                "description": "get sensor data",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "get sensor data by device id",
                        "required": true,
                        "schema": {
                            "properties": {
                                "devicesId": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/devices/getSensorDataChart": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "get sensor data for chart",
                "description": "get sensor data for chart",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "get sensor data by device id",
                        "required": true,
                        "schema": {
                            "properties": {
                                "devicesId": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Example Format: Bearer Your-Token",
            "name": "authorization",
            "in": "header"
        }
    }
}